--- static/Client.py
@@ -1,10 +1,0 @@
-class Client :
-    def __init__(self, first_name, last_name, dva_num, address, suburb, state, postcode, number):
-        self.first_name = first_name
-        self.last_name = last_name
-        self.dva_num = dva_num
-        self.address = address
-        self.suburb = suburb
-        self.state = state
-        self.postcode = postcode
-        self.number = number
